requirements.txt
Django>=2.1,<3.0
psycopg2-binary>=2.7,<3.0
pillow>=5.4,<6.0
djangorestframework>=3.9.1,<=3.9.2